feat: split search and results to different components
