+++ src/components/App/App.tsx
@@ -1,0 +1,46 @@
+import React, { useEffect, useState } from 'react';
+import styles from './App.module.scss';
+import { ResultItemsType, ResultResponse } from '../../types/types';
+import { getApiData } from '../../utils/api';
+import { Search } from '../Search/Search';
+import {
+  getLocalStorageSearchvalue,
+  setLocalStorageSearchValue,
+} from '../../utils/localStorage';
+import { Results } from '../Results/Results';
+
+export function App() {
+  const [searchResults, setSearchResults] = useState<ResultItemsType>([]);
+  const [isLoading, setIsLoading] = useState(false);
+
+  const searchHandler = (searchValue: string) => {
+    setIsLoading(true);
+    setLocalStorageSearchValue(searchValue);
+    getApiData(searchValue)
+      .then((response) => response.json())
+      .then((result: ResultResponse) => {
+        setIsLoading(false);
+        setSearchResults(result.results);
+      });
+  };
+
+  useEffect(() => {
+    setIsLoading(true);
+    const searchValue = getLocalStorageSearchvalue();
+    getApiData(searchValue)
+      .then((response) => response.json())
+      .then((result: ResultResponse) => {
+        setIsLoading(false);
+        setSearchResults(result.results);
+      });
+  }, []);
+
+  return (
+    <>
+      <main className={styles.main}>
+        <Search searchHandler={searchHandler} />
+        {isLoading ? 'Loading!!!' : <Results searchResults={searchResults} />}
+      </main>
+    </>
+  );
+}
