+++ src/components/ErrorButton/ErrorButton.module.scss
@@ -1,0 +1,6 @@
+.error_button {
+  border: 1px solid black;
+  border-radius: 5px;
+  width: 60px;
+  height: 20px;
+}
