+++ src/components/Results/Results.module.scss
@@ -1,0 +1,12 @@
+.result_block {
+  display: flex;
+  max-width: 70%;
+  margin: 10px auto;
+  justify-content: space-between;
+}
+
+.result_container {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+}
