+++ src/utils/localStorage.tsx
@@ -1,0 +1,8 @@
+export const setLocalStorageSearchValue = (searchValue: string) => {
+  localStorage.setItem('lev1ossa-react-components-value', searchValue);
+};
+
+export const getLocalStorageSearchvalue = () => {
+  const searchKeyWord = localStorage.getItem('lev1ossa-react-components-value');
+  return !!searchKeyWord ? searchKeyWord : '';
+};
